Split the factory methods into a trait.
